Finished a method to create a list of Fibo numbers with an index limit
Did you use Linter?
Yes

1. What exactly did you do, explain step by step using bullet points?

a. Added class Fibo with its first method createFibolist to create a list of Fibonacci numbers with an index limit.

2. Why did you do it?

To then limit the list to an value of less than 4 million for the last element of the list.

3. What is the result of your change/test?

Testing yields success.

4. Further comments: any hypothesis, future development suggestions, anything you learned?

It is very fast for the generation of Fibonacci numbers up to 100 elements. Takes just less then a second to do which is not what I expected.